requirements.txt
# AWS
boto3>=1.34.0
botocore>=1.34.0

# Backend
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
tiktoken>=0.5.0

# OpenSearch
opensearch-py>=2.4.0
requests-aws4auth>=1.2.3

# Frontend
streamlit>=1.31.0
requests>=2.31.0

# Utilities
hashlib-additional>=1.0.1
